% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iss.R
\name{iss_astronauts}
\alias{iss_astronauts}
\title{Fetch current astronauts in space}
\usage{
iss_astronauts()
}
\value{
tibble: name (character), craft (character)
}
\description{
Returns all people currently in space, including their spacecraft.
}
