% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iss.R
\name{iss_position}
\alias{iss_position}
\title{Fetch current ISS position}
\usage{
iss_position()
}
\value{
tibble: timestamp (POSIXct), latitude (numeric), longitude (numeric)
}
\description{
Returns the current latitude and longitude of the International Space Station.
}
